+++ packages/browser-wallet/src/popup/pages/Account/AccountDetails/AccountDetails.tsx
@@ -3,8 +3,8 @@ import React, { useEffect, useState } from 'react';
 import { displayAsCcd, getPublicAccountAmounts, PublicAccountAmounts } from 'wallet-common-helpers';
 import { useTranslation } from 'react-i18next';
 import { useAtomValue } from 'jotai';
-import { jsonRpcUrlAtom } from '@popup/store/settings';
 import { identityNamesAtom } from '@popup/store/identity';
+import { networkConfigurationAtom } from '@popup/store/settings';
 
 import { displaySplitAddress } from '@popup/shared/utils/account-helpers';
 import VerifiedIcon from '@assets/svg/verified-stamp.svg';
@@ -40,7 +40,7 @@ const zeroBalance: Omit<PublicAccountAmounts, 'scheduled'> = {
 
 export default function AccountDetails({ expanded, account, className }: Props) {
     const { t } = useTranslation('account', { keyPrefix: 'details' });
-    const jsonRpcUrl = useAtomValue(jsonRpcUrlAtom);
+    const { jsonRpcUrl } = useAtomValue(networkConfigurationAtom);
     const [balances, setBalances] = useState<Omit<PublicAccountAmounts, 'scheduled'>>(zeroBalance);
     const identityNames = useAtomValue(identityNamesAtom);