+++ packages/browser-wallet/src/popup/pages/ChangePasscode/i18n/da.ts
@@ -1,0 +1,13 @@
+import type en from './en';
+
+const t: typeof en = {
+    changePasscode: 'Opdater adgangskode',
+    incorrectPasscode: 'Forkert adgangskode',
+    labels: {
+        currentPasscode: 'Indtast nuværende adgangskode',
+        newPasscode: 'Indtast ny adgangskode',
+        newPasscodeRepeated: 'Gentag ny adgangskode',
+    },
+};
+
+export default t;