+++ packages/browser-wallet/src/popup/pages/ChangePasscode/i18n/en.ts
@@ -1,0 +1,11 @@
+const t = {
+    changePasscode: 'Change passcode',
+    incorrectPasscode: 'Incorrect passcode',
+    labels: {
+        currentPasscode: 'Enter current passcode',
+        newPasscode: 'Enter new passcode',
+        newPasscodeRepeated: 'Confirm new passcode',
+    },
+};
+
+export default t;