+++ packages/browser-wallet/src/shared/storage/access.ts
@@ -6,6 +6,7 @@ import {
     Theme,
     WalletCredential,
     IdentityProvider,
+    NetworkConfiguration,
 } from './types';
 
 export type StorageAccessor<V> = {
@@ -51,7 +52,10 @@ export const storedConnectedSites = makeStorageAccessor<Record<string, string[]>
     ChromeStorageKey.ConnectedSites
 );
 export const storedCredentials = makeStorageAccessor<WalletCredential[]>('local', ChromeStorageKey.Credentials);
-export const storedJsonRpcUrl = makeStorageAccessor<string>('local', ChromeStorageKey.JsonRpcUrl);
+export const storedCurrentNetwork = makeStorageAccessor<NetworkConfiguration>(
+    'local',
+    ChromeStorageKey.NetworkConfiguration
+);
 export const storedSelectedAccount = makeStorageAccessor<string>('local', ChromeStorageKey.SelectedAccount);
 export const storedEncryptedSeedPhrase = makeStorageAccessor<EncryptedData>('local', ChromeStorageKey.SeedPhrase);
 export const storedTheme = makeStorageAccessor<Theme>('local', ChromeStorageKey.Theme);