+++ packages/browser-wallet/CHANGELOG.md
@@ -5,6 +5,11 @@
 ### Added
 
 -   Login screen.
+-   Change passcode settings page.
+
+### Changed
+
+-   The network connections page has been updated with new UI.
 
 ## 0.3.0 2022-8-18